pyproject.toml
[project]
name = "energy-monitor"
version = "1.0.0"
description = "Energy meter monitoring system — fetches a prepaid electricity meter vendor API and notifies via Telegram"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.31.0",
    "matplotlib>=3.8.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.14",
    "freezegun>=1.5",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["scraper"]
